Refactor login flow
